Move language change logic into container
